Add new way to perform the profiling
The 'builder' package takes a specially formatted table description of the
network and generates a Torch network model and calculates a valid 'eye' size
for said network.

The 'profiler' package is able to take the same network description and
calculate the number of neurons, operations and run time performance. The
algorithms to do these calculations are the same as those used in the current
code.