Add FPGA 'nnx' as platform for profiling
The CLI arguments have been changed to allow for other types of platforms to
be selected, now it can do 'cpu' or 'cuda' or 'nnx'.

There is an assumption that the 'nnx-parser' repo can be found in the home
directory. This is needed to find nnx software used to do the profiling.