Remove references to CUDA/Tegra in spatial.lua
Spatial is used when the input image dimensions are larger than
those of the images used to train the network. This has nothing
to do with CUDA/Tegra. References to the CUDA API within spatial
were causing the profiler to crash on the Jetson dev board for
spatial modes 1 and 2.

These were removed. The network is converted to cuda through the
convertCUDNN function inside the profiler.