Bug fix: Kriz.lua definition
A layer is defined as split into columns if the "input" is split in 2 not if
the output is split. Also the "output" layer entry is currently the COMBINED
output of all columns, not the output on one column.

Using the original paper as the source of the numbers:

http://www.cs.toronto.edu/~fritz/absps/imagenet.pdf

Layer 1) This layer does not have a split input (as its RBG) and produces 96
         outputs.

Layer 2) The inputs are split into two groups of 48 maps. Each group is
	 processed separately to produce 128 outputs making a total of 256
	 outputs COMBINED. Thus it has 2 columns with an 256 output.

Layer 3) Is a mixing layer where the 256 inputs from layer 2 produce 384
         outputs. This layer is thus not split into columns.

Layer 4) The inputs are split into two groups of 192 maps. Each group is
	 processed separately to produce 192 outputs making a total of 384
	 outputs COMBINED. Thus it has 2 columns with an 384 output.

Layer 5) The inputs are split into two groups of 192 maps. Each group is
	 processed separately to produce 128 outputs making a total of 256
	 outputs COMBINED. Thus it has 2 columns with an 256 output.